Renamed http_utils module
